src/requirements.txt
Django>=2.0
djangorestframework
requests